requirements.txt
streamlit
transformers
opencv-python
requests
<<<<<<< HEAD
python-dotenv
=======
python-dotenv
torch
torchvision
torchaudio
>>>>>>> 83d2796264fd41c4247e5cac875f7d5ce6221eb2
